Prevent Gnome crash by failing gracefully when pushModal fails
The "Could not grab modal" error has been affecting my system running
Gnome 42 on Ubuntu 22.04 LTS for a few days now. Once the error is
reached and PaperWM raises this error, mouse input does not work across
Gnome on any window. I have to find an open Terminal window and somehow
disable the PaperWM extension in order to make mouse input work again.

I am not entirely sure what triggers the error, and I have not been able
to find a reliable way to trigger it. It often happens when I am rapidly
switching windows using keyboard shortcuts. The error comes from
PaperWM's "ActionDispatcher" class.

I believe that we should not "throw" an Error here, which crashes the
whole Gnome Shell. Instead, failing gracefully is a good first step:
With this change, the user will still be confused because their
"keyboard" shortcut action will not do anything. This is still better
than crashing Gnome Shell.

The next step should be to show an alert to the user which indicates to
them that their input could not be grabbed.