Check for all falsy values
I saw an error in the journal which shows that `dispatcher` was undefined. This would happen when
PaperWM is started immediately after Gnome Shell starts:

	Apr 07 09:36:43 skannan--20250106--BQ374 gnome-shell[3651]: JS ERROR: TypeError: dispatcher is undefined
																getActionDispatcher@/home/siddharth/.local/share/gnome-shell/extensions/paperwm@paperwm.github.com/navigator.js:470:2
																begin@/home/siddharth/.local/share/gnome-shell/extensions/paperwm@paperwm.github.com/grab.js:56:27
																grabBegin@/home/siddharth/.local/share/gnome-shell/extensions/paperwm@paperwm.github.com/tiling.js:3639:20
																init/<@/home/siddharth/.local/share/gnome-shell/extensions/paperwm@paperwm.github.com/tiling.js:1759:88

This was caused because I changed the check from `if (dispatcher)` which checks whether `dispatcher`
is truthy, to `if (dispatcher === null)`, which only checks if it is `null`. `undefined` would have
failed the previous check while the second check did succeed and ended up calling `dispatcher.mode
|= mode` which resulted in an error. I had to restart PaperWM in order to recover from this error.

This change restores the code back to checking for truthy values.

Also, all the log lines now include "grab modal" so that I can get a summarized view of how this
error is changing over time with the various attempted fixes.